Lizzy(bot)/requirements.txt
discord==2.0.0
discord.py==2.0.1
python-decouple==3.6
async-timeout==3.0.1